+++ core/include/karin/graphics.h
@@ -1,7 +1,7 @@
 #ifndef KARIN_GRAPHICS_H
 #define KARIN_GRAPHICS_H
 
-#include "graphics/WindowRenderer.h"
+#include "graphics/window_renderer.h"
 #include "graphics/graphics_context.h"
 #include "graphics/path.h"
 #include "graphics/stroke_style.h"
