+++ core/include/karin/graphics/window_renderer.h
